requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
dspy-ai>=2.0.0
streamlit-aggrid>=0.3.4
ollama>=0.1.0
chromadb>=0.4.22
tqdm>=4.65.0
anthropic>=0.5.0 